🐛

+++ src/app/api/delete-user/route.ts
@@ -25,37 +25,33 @@ export async function POST(req: NextRequest): Promise<NextResponse<ApiResponse>>
     error: sessionError,
   } = await supabase.auth.getSession();
 
-  if (sessionError) {
+  if (sessionError || !session) {
     console.error("Error:", sessionError);
-    return NextResponse.json({ error: sessionError.message }, { status: 500 });
+    return NextResponse.json({ error: "セッションが見つかりませんでした🥲" }, { status: 500 });
   }
 
-  if (session) {
-    try {
-      const user = await prisma.user.findUnique({
-        where: { id: session.user.id },
-        select: { profileImage: true },
-      });
+  try {
+    const user = await prisma.user.findUnique({
+      where: { id: session.user.id },
+      select: { profileImage: true },
+    });
 
-      const profilePath = getProfilePath(user?.profileImage ?? null);
+    const profilePath = getProfilePath(user?.profileImage ?? null);
 
-      // サーバー側のセッションを削除するためサインアウトさせる
-      await supabase.auth.signOut();
-      await supabaseClient.auth.admin.signOut(session.access_token);
+    // サーバー側のセッションを削除するためサインアウトさせる
+    await supabase.auth.signOut();
+    await supabaseClient.auth.admin.signOut(session.access_token);
 
-      await supabaseClient.auth.admin.deleteUser(session.user.id);
-      await prisma.user.delete({ where: { id: session.user.id } });
-      if (profilePath) {
-        await supabaseClient.storage.from("user").remove([profilePath]);
-      }
-
-      return NextResponse.json({ message: "退会しました🔥" }, { status: 200 });
-    } catch (error) {
-      console.error("Error during user deletion:", error);
-      return NextResponse.json({ error: "退会に失敗しました🥲" }, { status: 500 });
+    await supabaseClient.auth.admin.deleteUser(session.user.id);
+    await prisma.user.delete({ where: { id: session.user.id } });
+    if (profilePath) {
+      await supabaseClient.storage.from("user").remove([profilePath]);
     }
-  } else {
-    return NextResponse.json({ error: "セッションが見つかりませんでした🥲" }, { status: 500 });
+
+    return NextResponse.json({ message: "退会しました🔥" }, { status: 200 });
+  } catch (error) {
+    console.error("Error during user deletion:", error);
+    return NextResponse.json({ error: "退会に失敗しました🥲" }, { status: 500 });
   }
 }
